+++ packages/bigby-cli/src/commands/index.ts
@@ -1,0 +1,4 @@
+import dev from "./dev";
+import build from "./build";
+
+export default { dev, build };