+++ packages/bigby-cli/src/webpack/configuration.ts
@@ -1,0 +1,77 @@
+import path from "path";
+import webpack from "webpack";
+
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const CopyPlugin = require("copy-webpack-plugin");
+const { CleanWebpackPlugin } = require("clean-webpack-plugin");
+const WebpackBar = require("webpackbar");
+const BundleAnalyzerPlugin = require("webpack-bundle-analyzer")
+  .BundleAnalyzerPlugin;
+import fs from "fs";
+
+export const create = () => {
+  const cwd = process.cwd();
+
+  /* Which files and directories should we copy over? */
+  const directoryExists = (name: string) =>
+    fs.existsSync(path.resolve(cwd, name));
+
+  let copyObjects = new Array<any>();
+  if (directoryExists("assets"))
+    copyObjects.push({ from: "assets", to: "assets" });
+
+  /* Return a webpack configuration object */
+  return <webpack.Configuration>{
+    context: path.resolve(cwd),
+    entry: path.resolve(__dirname, "../entrypoint.js"),
+    mode: "development",
+    devtool: "inline-source-map",
+    output: {
+      path: path.resolve(cwd, "./dist"),
+      filename: "[name].bundle.js",
+    },
+    module: {
+      rules: [
+        { test: /\.tsx?$/, use: "ts-loader" },
+        { test: /\.css$/, use: ["style-loader", "css-loader"] },
+      ],
+    },
+    resolve: {
+      extensions: [".tsx", ".ts", ".js"],
+      alias: {
+        "@": cwd,
+      },
+    },
+    plugins: [
+      new WebpackBar(),
+      new CleanWebpackPlugin(),
+      new CopyPlugin(copyObjects),
+      new HtmlWebpackPlugin({
+        title: "Bigby Game",
+        template: require("html-webpack-template"),
+        appMountId: "app",
+      }),
+    ],
+    optimization: {
+      splitChunks: {
+        chunks: "all",
+      },
+    },
+    performance: {
+      maxEntrypointSize: 1000000,
+      maxAssetSize: 1000000,
+    },
+  };
+};
+
+export const setProduction = (
+  config: webpack.Configuration,
+  production = false
+) => {
+  config.mode = production ? "production" : "development";
+  config.devtool = production ? false : "inline-source-map";
+};
+
+export const enableAnalyzer = (config: webpack.Configuration) => {
+  config.plugins.push(new BundleAnalyzerPlugin());
+};