+++ packages/bigby/src/index.ts
@@ -1,0 +1,4 @@
+import { Entity, Behavior } from "@bigby/core";
+
+export { Entity };
+export { Behavior };