+++ CHANGELOG.md
@@ -9,7 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- @bigby/core: `Entity.from` factory that will create a complete entity (including behaviors) from data.
+- @bigby/cli: new module, providing a convenient `bigby` CLI tool.
 - @bigby/editor: provide `startEditor` convenience function
+
+### Changed
+
+- @bigby/editor: now automatically installs React as a dependency.
 
 ## 0.5.10